template/requirements.txt
# Local package
-e .

# Research and data analysis packages
matplotlib
numpy
pandas
scikit-learn
scipy
seaborn

# Data management
dvc
fastds

# Experiment management
mlflow

# Documentation
Sphinx

# Interactive environments
ipython
jupyterlab

# Basic Python utility packages
python-dotenv>=0.5.1
PyYAML
python-slugify
typer

# Code quality
coverage
flake8
pycodestyle
pylint
pytest
pytest-cov
pytest-pycodestyle
pytest-pylint
pytest-xdist
radon
